requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.10.0
aniso8601==9.0.1
annotated-types==0.5.0
antiorm==1.2.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
avro==1.11.1
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==39.0.1
Deprecated==1.2.13
factory-boy==3.2.1
Faker==17.6.0
flake8==6.0.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Limiter==3.3.0
Flask-Login==0.6.2
Flask-RESTful==0.3.9
flask-restful-swagger-3==0.4.9
Flask-SQLAlchemy==3.0.3
flask-swagger-ui==4.11.1
frozenlist==1.3.3
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.5.0
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
mccabe==0.7.0
mdurl==0.1.2
mock==5.1.0
multidict==6.0.4
mypy-extensions==1.0.0
ordered-set==4.1.0
orjson==3.8.6
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pluggy==0.13.1
psycopg2==2.9.5
py==1.11.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.5
pydantic_core==2.1.2
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.6.0
pytest==6.1.2
pytest-asyncio==0.20.1
pytest-metadata==2.0.4
pytest-mock==3.11.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
redis==4.2.0rc1
requests==2.28.2
rich==13.4.2
six==1.16.0
spectree==1.0.3
SQLAlchemy==2.0.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.6.0
urllib3==1.26.14
uvicorn==0.20.0
Werkzeug==2.2.3
wrapt==1.15.0
yarl==1.8.2
zipp==3.14.0
zope.event==4.6
zope.interface==5.5.2
